speed up glove training (by factor of 3x): 1) use float where possible 2) use single representation for main and context word - this significantly improves convergence speed on small datasets

+++ src/utils.cpp
@@ -3,7 +3,7 @@
 using namespace Rcpp;
 using namespace std;
 
-NumericMatrix convert2Rmat(vector<vector<double> > &mat, size_t ncol) {
+NumericMatrix convert2Rmat(vector<vector<float> > &mat, size_t ncol) {
   NumericMatrix res(mat.size(), ncol);
   for (size_t i = 0; i < mat.size(); i++)
     for (size_t j = 0; j < ncol; j++)
@@ -19,23 +19,23 @@ NumericMatrix convert2Rmat(vector<vector<double> > &mat, size_t ncol) {
 //   return res;
 // }
 
-void fill_mat_val(vector<vector<double> > &mat, size_t ncol, double val) {
+void fill_mat_val(vector<vector<float> > &mat, size_t ncol, float val) {
   for (size_t i = 0; i < mat.size(); i++)
     for (size_t j = 0; j < ncol; j++)
       mat[i][j] = val;
 }
-void fill_mat_rand(vector<vector<double> > &mat, size_t ncol, double runif_min, double runif_max) {
+void fill_mat_rand(vector<vector<float> > &mat, size_t ncol, float runif_min, float runif_max) {
   for (size_t i = 0; i < mat.size(); i++)
     for (size_t j = 0; j < ncol; j++)
       mat[i][j] = R::runif(runif_min, runif_max); //(double)rand() / (double)RAND_MAX - 0.5;
 }
 
-void fill_vec_rand(vector<double>  &vec, double runif_min, double runif_max) {
+void fill_vec_rand(vector<float>  &vec, float runif_min, float runif_max) {
   for (size_t i = 0; i < vec.size(); i++)
     vec[i] = R::runif(runif_min, runif_max); //(double)rand() / (double)RAND_MAX - 0.5;
 }
 
-void fill_vec_val(vector<double>  &vec, double val) {
+void fill_vec_val(vector<float>  &vec, float val) {
   for (size_t i = 0; i < vec.size(); i++)
     vec[i] = val;
 }
